ensure Array.shape is a simplified Array or an int
This fixes the following snippet.

    from nutils import mesh, function

    domain, geom = mesh.rectilinear([[-1, 0, 1]]*2)
    domain = domain.trim(function.norm2(geom)-.5, maxrefine=0)
    function.chain([domain.basis('spline', degree=2), [1.]])

    # initialize namespace
    ns = function.Namespace()
    ns.xybasis = function.zeros([16])
    ns.phi = 'xybasis_n ?lhs_n'

The length of the first axis of `ns.xybasis` was a `numpy.int64`.  This is
unfortunately `==` to `(16,)` or `nutils.expression._Length(16)`, which
identifies the index `'n'` of `'lhs_n'`, and the unknown length is therefore
removed from the list of unknown lengths.  Hence the error

    nutils.expression.ExpressionSyntaxError: Length of axis cannot be determined from the expression.
    xybasis_n ?lhs_n
                    ^